Move App component to src directory and colocate tests
