+++ src/App/App.spec.js
